requirements.txt
google-generativeai==0.8.1
pillow==10.4.0
streamlit==1.38.0
streamlit-option-menu==0.3.13

